Lab 1
Subject area: Housing and Utilities Office. In the class, store information about the district to which the housing and utility office belongs, the housing and utility office number, the number of residents, the monthly payment (one for all residents), and the number of those who paid. Implement a method for calculating the total debt of residents.